enkrypt-in-agent-sdk/pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "enkrypt-agent-sdk"
version = "0.1.0"
description = "Enkrypt AI In-Agent Security SDK — embed guardrails, observability, and PII protection directly inside AI agents."
readme = "README.md"
requires-python = ">=3.10"
license = "Apache-2.0"
authors = [
    { name = "Enkrypt AI", email = "support@enkryptai.com" },
]
keywords = ["ai", "security", "guardrails", "agents", "opentelemetry", "mcp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "aiohttp>=3.8.0",
]

[project.optional-dependencies]
otel = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.20.0",
    "opentelemetry-exporter-otlp-proto-http>=1.20.0",
]
langchain = [
    "langchain-core>=0.1.0",
]
openai = [
    "openai>=1.0.0",
    "openai-agents>=0.1.0",
]
anthropic = [
    "anthropic>=0.30.0",
]
all = [
    "enkrypt-agent-sdk[otel,langchain,openai,anthropic]",
]
dev = [
    "enkrypt-agent-sdk[all]",
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.4.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
